pequenas mudanças
fiz algumas mudanças no codigo pra poder ser explicado na aula de amanha
