--- .env
@@ -1,3 +1,0 @@
-PORT=4000
-DB_USER=user123
-DB_PASS=VHMuUzQOAkArLOP4
