+++ config/multer.js
@@ -1,0 +1,17 @@
+const multer = require('multer');
+const path = require('path');
+
+/*
+const storage = multer.diskStorage({
+    destination: function(req,res, cb) {
+        cb();
+    },                   
+    filename: function(req, file, cb) {
+        cb();
+    },
+});
+*/
+
+const upload;
+
+module.exports = upload;
