+++ routes/image.js
@@ -1,0 +1,13 @@
+const express = require('express');
+const router = express.Router();
+
+const upload = require("../config/multer")
+
+const ImageController = require("../controller/imageController");
+
+/*
+router.post();
+router.get("/", ImageController.getAll);
+*/
+
+module.exports = router;
