+++ sample.env
@@ -1,0 +1,3 @@
+PORT=4000
+DB_USER=
+DB_PASS=
